% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_sat.R
\docType{data}
\name{d_sat}
\alias{d_sat}
\title{Saturation curve data}
\usage{
data(d_sat)
}
\description{
Sample saturation curve data featuring time, conc. of substrate, replicate #, and spectral data.
}
\author{
Christopher L. Cook
}
