% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_sat_g.R
\docType{data}
\name{d_sat_g}
\alias{d_sat_g}
\title{Saturation curve data}
\usage{
data(d_sat_g)
}
\description{
Sample saturation curve data featuring time, conc. of substrate, replicate #, spectral data, assay volume, substrate control, buffer volume, homogenate volume, and soil mass.
}
\author{
Christopher L. Cook and Andrew D. Steen
}
