% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_sat_n.R
\docType{data}
\name{d_sat_n}
\alias{d_sat_n}
\title{Saturation curve data, normalized}
\usage{
data(d_sat_n)
}
\description{
Saturation curve data featuring time, conc. of substrate, replicate #, spectral data, and normalization factor.
}
\author{
Christopher L. Cook
}
