% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_std_g.R
\docType{data}
\name{d_std_g}
\alias{d_std_g}
\title{Standard curve data}
\usage{
data(d_std_g)
}
\description{
Sample standard curve data featuring spectral data in presence of homogenate, spectral data in presence of buffer, and conc. of standard.
}
\author{
Christopher L. Cook and Andrew D. Steen
}
