% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmmek-package.R
\docType{package}
\name{ezmmek-package}
\alias{ezmmek}
\alias{ezmmek-package}
\title{ezmmek: Easy Michaelis-Menten Enzyme Kinetics}
\description{
Analysis of enzyme activities on fluorogenic and chromogenic substrates using Michaelis-Menten kinetics. It outputs plots and relevant statistics regarding standard curves and saturation curves. It also enables the user to customize plots to their particular datasets. 
    This package is based on concepts described in: German et al. (2011) <doi:10.1016/j.soilbio.2011.03.017>, Sinsabaugh et al. (2014) <doi:10.1007/s10533-014-0030-y>, and Steen and Arnosti (2011) <doi:10.1016/j.marchem.2010.10.006>.
}
\keyword{internal}
