% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmmek_calc_mm_fit.R
\name{ezmmek_calc_mm_fit}
\alias{ezmmek_calc_mm_fit}
\title{ezmmek_calc_mm_fit}
\usage{
ezmmek_calc_mm_fit(df, km, vmax)
}
\arguments{
\item{df}{Dataframe of class 'new_ezmmek_calibrate'}

\item{km}{Starting value to estimate km. Default value is median of 'sub.conc' values}

\item{vmax}{Starting value to estimate vmax. Default value is max activity calculated}
}
\description{
Calculate Michaelis-Menten fit
}
