% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmmek_calc_std_lm_buffer.R
\name{ezmmek_calc_std_lm_buffer}
\alias{ezmmek_calc_std_lm_buffer}
\title{ezmmek_calc_std_lm_homo}
\usage{
ezmmek_calc_std_lm_buffer(df)
}
\arguments{
\item{df}{Standard curve dataframe}
}
\description{
Calculate linear model for standard curve in homogenate
}
