% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmmek_calc_std_lm_homo.R
\name{ezmmek_calc_std_lm_homo}
\alias{ezmmek_calc_std_lm_homo}
\title{ezmmek_calc_std_lm_homo}
\usage{
ezmmek_calc_std_lm_homo(df)
}
\arguments{
\item{df}{Standard curve dataframe}
}
\description{
Calculate linear model for standard curve in homogenate
}
