% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmmek_calc_std_lm_homo_buf.R
\name{ezmmek_calc_std_lm_homo_buffer}
\alias{ezmmek_calc_std_lm_homo_buffer}
\title{ezmmek_calc_std_lm_homo_buffer}
\usage{
ezmmek_calc_std_lm_homo_buffer(df)
}
\arguments{
\item{df}{Standard curve dataframe}
}
\description{
Calculate linear model for standard curve in homogenate-buffer solution
}
