% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmmek_calibrate_activities.R
\name{ezmmek_calibrate_activities}
\alias{ezmmek_calibrate_activities}
\title{ezmmek_calibrate_activities}
\usage{
ezmmek_calibrate_activities(df, method, columns)
}
\arguments{
\item{df}{Joined dataframes of class 'new_ezmmek_std_group' and 'new_ezmmek_act_group'}

\item{method}{Enzyme assay protocol. Must define method as 'isc' or 'ibc'}

\item{columns}{Column names carried from parent functions}
}
\description{
Calibrates enzyme activity data by standard curve
}
