% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmmek_std_lm.R
\name{ezmmek_std_lm}
\alias{ezmmek_std_lm}
\title{ezmmek_std_lm}
\usage{
ezmmek_std_lm(df, method = method, columns = NULL)
}
\arguments{
\item{df}{Standard curve dataframe}

\item{method}{Enzyme assay protocol. Must define method as 'steen' or 'german'}

\item{columns}{Column names carried over from parent functions if parent functions used}
}
\description{
Creates dataframe of standard curve models as determined by grouping of user-defined columns
}
