% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_std_curve.R
\name{g_std_curve}
\alias{g_std_curve}
\title{Create standard curves (separate quench)}
\usage{
g_std_curve(d_std_g, x.label = NULL, y.label = NULL)
}
\arguments{
\item{d_std_g}{Must be a dataframe that contains 'std.conc', 'replicate', 'spec.homo', and 'spec.buffer'.}

\item{x.label}{Input a character string to label the x-axis.}

\item{y.label}{Input a character string to label the y-axis.}
}
\value{
List containing plots, fit models, and new dataframe.
}
\description{
Creates a standard curve of spectral data versus fluorophore/chromophore concentration.
Linear models and standard curves are generated for the standard in the presence of buffer and homogenate.
A third plot that overlays both standard curves is also generated.
}
\details{
Plots spectral data vs standard concentration.
Can label the x-axis by inputting an argument.
Can label the y-axis by inputting an argument.
If these arguments are not defined, the function will default to create a plot without labels.
}
\examples{
g_std_curve(d_std_g)
g_std_curve(d_std_g, x.label = "x-axis label", y.label = "y-axis label")

}
\author{
Christopher L. Cook and Andrew D. Steen
}
