% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ezmmek_std_group.R
\name{new_ezmmek_std_group}
\alias{new_ezmmek_std_group}
\title{new_ezmmek_std_group}
\usage{
new_ezmmek_std_group(std.data.fn, ..., method = NA, columns = NULL)
}
\arguments{
\item{std.data.fn}{Standard data file as character string}

\item{...}{User defined column names to group std.data.fn}

\item{method}{Enzyme assay protocol. Must define method as 'steen' or 'german'}

\item{columns}{Column names carried over from parent functions if parent functions used}
}
\description{
Groups standard curve data by user-defined columns
}
\examples{
new_ezmmek_std_group(std.data.fn, site.name, std.type, method = "steen", columns = NULL)
}
