% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_new_ezmmek_act_group.R
\name{plot.new_ezmmek_act_group}
\alias{plot.new_ezmmek_act_group}
\title{plot_new_ezmmek_act_group}
\usage{
\method{plot}{new_ezmmek_act_group}(x, ...)
}
\arguments{
\item{x}{data.frame object of class new_ezmmek_act_group}

\item{...}{User defined column names by which to facet plot}
}
\description{
Plots new_ezmmek_act_group object and facets by specified column names
}
\examples{
\dontrun{plot.new_ezmmek_act_group(new_ezmmek_act_group_obj,
site_name,
std_type)}
}
