% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_new_ezmmek_calibrate.R
\name{plot.new_ezmmek_calibrate}
\alias{plot.new_ezmmek_calibrate}
\title{plot_new_ezmmek_calibrate}
\usage{
\method{plot}{new_ezmmek_calibrate}(x, ...)
}
\arguments{
\item{x}{data.frame object of class new_ezmmek_calibrate}

\item{...}{User defined column names by which to facet plot}
}
\description{
Plots new_ezmmek_calibrate object and facets by specified column names
}
\examples{
\dontrun{plot.new_ezmmek_calibrate(new_ezmmek_calibrate_obj,
site_name,
std_type)}
}
