% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_new_ezmmek_sat_fit.R
\name{plot.new_ezmmek_sat_fit}
\alias{plot.new_ezmmek_sat_fit}
\title{plot_new_ezmmek_sat_fit}
\usage{
\method{plot}{new_ezmmek_sat_fit}(x, ...)
}
\arguments{
\item{x}{data.frame object of class new_ezmmek_sat_fit}

\item{...}{User defined column names by which to facet plot}
}
\description{
Plots new_ezmmek_sat_fit object and facets by specified column names
}
\examples{
\dontrun{plot.new_ezmmek_act_group(new_ezmmek_sat_fit_obj,
site_name,
stdy_type)}
}
