% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_sat_curve.R
\name{s_sat_curve}
\alias{s_sat_curve}
\title{Create saturation curve (bulk quench)}
\usage{
s_sat_curve(d_std, d_sat, x.label = NULL, y.label = NULL, km = NULL,
  vmax = NULL)
}
\arguments{
\item{d_std}{Must be a dataframe that contains 'std.conc' and 'spec'.}

\item{d_sat}{Must be a dataframe that contains 'time', 'sub.conc' (substrate concentration), 'replicate', and 'spec' (spectral data).}

\item{x.label}{Input a character string to label the x-axis. Default is 'NULL'.}

\item{y.label}{Input a character string to label the y-axis. Default is 'NULL'.}

\item{km}{Input a starting value to estimate 'km' value. Default value is median of 'sub.conc' values.}

\item{vmax}{Input a starting value to estimate 'vmax' value. Default value is max activity ('slope.m') calculated.}
}
\value{
List containing new dataframe, regression model, and saturation curve.
}
\description{
Creates a dataframe and plot by applying the standard curve coefficients to the raw saturation data.
}
\details{
The spectral data is converted to concentration of standard.
The new dataframe contains the average slope (rate of reaction) and standard deviation for each replicate at each substrate concentration.
'sat_curve' plots the new dataframe with substrate concentration on the x-axis, and rate of reaction on the y-axis.
It predicts and reports Vmax and Km values.
It creates a list output containing the new dataframe, an additional new dataframe consisting of predicted curve fit values, the regression model, and the saturation curve plot.
}
\examples{
s_sat_curve(d_std, d_sat)
s_sat_curve(d_std, d_sat, x.label = "x-axis label", y.label = "y-axis label", km = 97, vmax = 0.03)

}
\author{
Christopher L. Cook and Andrew D. Steen
}
