% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_std_curve.R
\name{s_std_curve}
\alias{s_std_curve}
\title{Create standard curve (bulk quench)}
\usage{
s_std_curve(d_std, x.label = NULL, y.label = NULL)
}
\arguments{
\item{d_std}{Must be a dataframe that contains 'std.conc', 'spec', and 'replicate'.}

\item{x.label}{Input a character string to label the x-axis.}

\item{y.label}{Input a character string to label the y-axis.}
}
\value{
List containing plot, fit model, and new dataframe.
}
\description{
Creates a standard curve of spectral data versus fluorophore/chromophore concentration.
}
\details{
Plots spectral data vs standard concentration.
Can label the x-axis by inputting an argument.
Can label the y-axis by inputting an argument.
If these arguments are not defined, the function will default to create a plot without labels.
}
\examples{
s_std_curve(d_std)
s_std_curve(d_std, x.label = "x-axis label", y.label = "y-axis label")

}
\author{
Christopher L. Cook and Andrew D. Steen
}
