% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_curve.R
\name{sat_curve}
\alias{sat_curve}
\title{Create saturation curve}
\usage{
sat_curve(d_std, d_sat, x.label = NULL, y.label = NULL)
}
\arguments{
\item{d_std}{Must be a dataframe that contains 'std.conc' and 'spec'.}

\item{d_sat}{Must be a dataframe that contains 'time', 'sub.conc' (substrate concentration), 'replicate', and 'spec' (spectral data).}

\item{x.label}{The user must input a character string in the third argument to label the x-axis.}

\item{y.label}{The user must input a character string in the fourth argument to label the y-axis.}
}
\value{
List containing new dataframe, regression model, and saturation curve.
}
\description{
Creates a dataframe and plot by applying the standard curve coefficients to the raw saturation data.
}
\details{
The spectral data is converted to concentration of standard.
The new dataframe contains the average slope (rate of reaction) and standard deviation for each replicate at each substrate concentration.
'sat_curve' plots the new dataframe with substrate concentration on the x-axis, and rate of reaction on the y-axis.
It predicts and reports Vmax and Km values.
It creates a list output containing the new dataframe, an additional new dataframe consisting of predicted curve fit values, the regression model, and the saturation curve plot.
}
\examples{
sat_curve(d_std, d_sat)
sat_curve(d_std, d_sat, "x-axis", "y-axis")

}
\author{
Christopher L. Cook and Andrew D. Steen
}
