% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_raw.R
\name{sat_raw}
\alias{sat_raw}
\title{Create plot of raw data}
\usage{
sat_raw(d_sat, x.label = NULL, y.label = NULL)
}
\arguments{
\item{d_sat}{Must be a dataframe that contains 'time', 'sub.conc' (substrate concentration), 'replicate', and 'spec' (spectral data).}

\item{x.label}{The user must input a character string in the second argument to label the x-axis.}

\item{y.label}{The user must input a character string in the third argument to label the y-axis.}
}
\value{
List containing plot.
}
\description{
Creates plot of raw saturation data.
}
\details{
Plots raw saturation curve data into separate facets based on substrate concentration ('sub.conc').
The user can label the x-axis by inputting a second argument.
The user can label the y-axis by inputting a third argument.
If the user does not define these arguments, the function will default to create a plot without labels.
}
\examples{
sat_raw(d_sat)
sat_raw(d_sat, "x-axis", "y-axis")

}
\author{
Christopher L. Cook and Andrew D. Steen
}
