% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_curve.R
\name{std_curve}
\alias{std_curve}
\title{Create standard curve}
\usage{
std_curve(d_std, x.label = NULL, y.label = NULL)
}
\arguments{
\item{d_std}{Must be a dataframe that contains 'std.conc' and 'spec'.}

\item{x.label}{The user must input a character string in the second argument to label the x-axis.}

\item{y.label}{The user must input a character string in the third argument to label the y-axis.}
}
\value{
List containing plot, fit model, and new dataframe.
}
\description{
Creates a standard curve of spectral data versus fluorophore/chromophore concentration.
}
\details{
Plots spectral data vs standard concentration.
The user can label the x-axis by inputting a second argument.
The user can label the y-axis by inputting a third argument.
If the user does not define these arguments, the function will default to create a plot without labels.
}
\examples{
std_curve(d_std)
std_curve(d_std, "x-axis", "y-axis")

}
\author{
Christopher L. Cook and Andrew D. Steen
}
