% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmmek-package.R
\docType{package}
\name{ezmmek-package}
\alias{ezmmek}
\alias{ezmmek-package}
\title{ezmmek: Analysis of Enzyme Activities on Synthetic Substrates using Michaelis-Menten Kinetics}
\description{
Analysis of enzyme activities on fluorogenic and chromogenic substrates using Michaelis-Menten kinetics. It outputs plots and relevant statistics regarding standard curves and saturation curves. It also enables the user to customize plots to their particular datasets.
}
\keyword{internal}
