% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ezmmek_calibrate.R
\name{new_ezmmek_act_calibrate}
\alias{new_ezmmek_act_calibrate}
\title{new_ezmmek_sat_fit}
\usage{
new_ezmmek_act_calibrate(std.data.fn, act.data.fn, ..., method = NA,
  conc.units = NA, signal.units = NA, columns = NULL)
}
\arguments{
\item{std.data.fn}{Standard data file as character string}

\item{act.data.fn}{Activity data file as character string}

\item{...}{User defined column names to join std.data.fn and act.data.fn}

\item{method}{Enzyme assay protocol. Must define method as 'steen' or 'german'}
}
\description{
Creates dataframe containing calibrated enzyme activity data of class 'new_ezmmek_calibrate'
}
\examples{
new_ezmmek_act_calibrate(std.data.fn, act.data.fn, site.name, std.type, km = NULL, vmax = NULL, method = "steen")
}
