% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sat_curve.R
\name{p_sat_curve}
\alias{p_sat_curve}
\title{Create saturation curve}
\usage{
p_sat_curve(d_std, d_sat)
}
\arguments{
\item{d_std}{Must be a dataframe that contains 'std.conc' and 'spec'.}

\item{d_sat}{Must be a dataframe that contains 'time', 'sub.conc' (substrate concentration), 'replicate', and 'spec' (spectral data).
If d_sat contains a fifth column, that fifth column will be assumed to be a normalization factor.
The rate of reacation will be divided by the values present in the fifth column.
The user will be prompted to name the unit of normalization, which will appear on the y-axis, if a fifth column is present.}
}
\value{
List containing new dataframe, regression model, and saturation curve.
}
\description{
Creates a dataframe and plot by applying the standard curve coefficients to the raw saturation data.
}
\details{
The spectral data is converted to concentration of standard.
The new dataframe contains the average slope (rate of reaction) and standard deviation for each replicate at each substrate concentration.
'p_sat_curve' plots the new dataframe with substrate concentration on the x-axis, and rate of reaction on the y-axis.
It asks the user to specify axis labels with the appropriate units. It predicts and reports Vmax and Km values.
It creates a list output containing the new dataframe, an additional new dataframe consisting of predicted curve fit values, the regression model, and the saturation curve plot.
}
\author{
Christopher L. Cook and Andrew D. Steen
}
