% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sat_raw.R
\name{p_sat_raw}
\alias{p_sat_raw}
\title{Create plot of raw data}
\usage{
p_sat_raw(d_sat)
}
\arguments{
\item{d_sat}{Must be a dataframe that contains 'time', 'sub.conc' (substrate concentration), 'replicate', and 'spec' (spectral data).}
}
\value{
List containing plot
}
\description{
Creates plot of raw saturation data
}
\details{
Plots raw saturation curve data into separate facets based on substrate concentration ('sub.conc').
It asks the user to specify the axis labels with the appropriate units.
It also creates a list output containing the raw data plot.
}
\examples{
p_sat_raw(d_sat)
p_sat_raw(d_sat_n)
}
\author{
Christopher L. Cook and Andrew D. Steen
}
