% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_std_curve.R
\name{p_std_curve}
\alias{p_std_curve}
\title{Create standard curve}
\usage{
p_std_curve(d_std)
}
\arguments{
\item{d_std}{Must be a dataframe that contains 'std.conc' and 'spec'.}
}
\value{
List containing plot and fit model
}
\description{
Creates a standard curve of spectral data versus fluorophore/chromophore concentration
}
\details{
Plots spectral data vs standard concentration.
It asks the user to specify the axis labels with the appropriate units.
It reports linear model summary statistics.
It also creates a list output containing the summary statistics and standard curve plot.
}
\author{
Christopher L. Cook and Andrew D. Steen
}
